Corrections (#8)
* Fix some typos and the path for config properties file.

Maybe an extra task for participants should be added at the end to use Configuration instead of URL's defined in constants in all projects in the workshop.

This will make consistent the code snippets used later in FT chapter.

* Small correction

* Small correction

* Fix Prometeus->Prometheus format name. Fix some typos and change ContentRepository -> ArticleRepository. Fix numbering for the steps at the end of the chapter.
Use recommended method to register metric instead of the deprecated one.

* Add Cancel button

* Add Back link to return to previous page.

* Here is the additional exercise

* This is needed in case someone decided to remove the defaultAuthor method after adding the fallback class.
In case fallback method is missing during the build open liberty fails but server process remains live and locks the target directory. You have to find the process in order to kill it.